+++ CHANGELOG.md
@@ -8,7 +8,7 @@ This document is formatted according to the principles of [Keep A CHANGELOG](htt
 
 ---
 
-## [Unreleased]
+## [v0.12.4]
 
 ### Added
 
@@ -230,6 +230,7 @@ This document is formatted according to the principles of [Keep A CHANGELOG](htt
 
 <!-- Releases -->
 
+[v0.12.4]: https://github.com/cucumber/godog/compare/v0.12.3...v0.12.4
 [v0.12.3]: https://github.com/cucumber/godog/compare/v0.12.2...v0.12.3
 [v0.12.2]: https://github.com/cucumber/godog/compare/v0.12.1...v0.12.2
 [v0.12.1]: https://github.com/cucumber/godog/compare/v0.12.0...v0.12.1